requirements.txt
Django
django-bootstrap4
whitenoise
djangorestframework
pylint
django-import-export
